requirements.txt
# Production Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
tiktoken>=0.5.0
sqlparse>=0.4.4
redis>=5.0.0
chromadb>=0.4.0
sentence-transformers>=2.3.0
python-dotenv>=1.0.0
pandas>=2.2.0
tabulate>=0.9.0
pyyaml>=6.0.1
structlog>=24.1.0
prometheus-client>=0.19.0

# Development Dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
ruff>=0.1.0
ipython>=8.20.0
ipdb>=0.13.13
